+++ app/src/main/java/uz/glight/hobee/distribuition/room/entity/SavedMedEntity.kt
@@ -13,4 +13,4 @@ data class SavedMedEntity(
     var name: String,
     var priceForOne: String,
     var amount: Int,
-    var allPrice: Double)
+    var allPrice: Long)
